+++ apps/storybook/package.json
@@ -17,6 +17,9 @@
 	},
 	"dependencies": {
 		"@ethereumjs/rlp": "^5.0.2",
+		"@geist/domain": "workspace:*",
+		"@geist/graphql": "workspace:*",
+		"@geist/ui-react": "workspace:*",
 		"@hookform/resolvers": "^3.9.0",
 		"@ipld/dag-ucan": "^3.4.0",
 		"@noble/ed25519": "2.1.0",
@@ -56,9 +59,6 @@
 		"@radix-ui/react-toggle": "^1.1.0",
 		"@radix-ui/react-toggle-group": "^1.0.4",
 		"@radix-ui/react-tooltip": "^1.1.2",
-		"@geist/domain": "workspace:*",
-		"@geist/graphql": "workspace:*",
-		"@geist/ui-react": "workspace:*",
 		"@scure/base": "^1.1.9",
 		"@storybook/components": "^8.3.6",
 		"@storybook/icons": "^1.2.12",
@@ -84,6 +84,7 @@
 		"lucide-react": "^0.446.0",
 		"micro-packed": "^0.6.3",
 		"multiformats": "^13.3.0",
+		"nanostores": "0.11.3",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-dropzone": "^14.2.9",
