--- apps/storybook/src/stories/signature/SignatureForm.mdx
@@ -1,28 +1,0 @@
-import { Meta, Story, Canvas } from '@storybook/blocks';
-import * as SignatureFormStories from './SignatureForm.stories';
-import * as SignatureFormMinStories from './SignatureFormMin.stories';
-
-
-<Meta title="MDX/Avatar" of={SignatureFormStories} />
-
-# Signature
-
-This is Demo of signing plain text message with [ERC-191](https://eips.ethereum.org/EIPS/eip-191)
-
-
-### With wagmi
-
-- it uses viem
-
-<Canvas>
-  <Story of={SignatureFormStories.Ens} />
-</Canvas>
-
-
-### With micro-eth-Signature
-- minimized dependency
-
-<Canvas>
-  <Story of={SignatureFormMinStories.EIP712Minimal} />
-</Canvas>
-
